+++ shared/src/commonMain/kotlin/no/synth/where/data/db/Migrations.kt
@@ -15,3 +15,9 @@ val MIGRATION_2_3: Migration = object : Migration(2, 3) {
         connection.execSQL("ALTER TABLE tracks ADD COLUMN sourceId TEXT")
     }
 }
+
+val MIGRATION_3_4: Migration = object : Migration(3, 4) {
+    override fun migrate(connection: SQLiteConnection) {
+        connection.execSQL("ALTER TABLE tracks ADD COLUMN color TEXT")
+    }
+}